refactor(import): read STEP assemblies in a sandbox; drop the partcad[cad] extra
'pc import assembly' and 'pc convert' now work with no CAD library in the core
process, and PartCAD declares no CAD dependency at all - not even an optional
extra.

- import assembly: the STEP-CAF reader, XDE-tree walk, per-part zeroing/dedup and
  STEP writing move into a new sandbox wrapper (wrappers/wrapper_import_assy.py).
  The core actions/assembly/import_assy.py drives it, registers the parts and
  writes the .assy - it never touches a live OCP object. The .assy it produces is
  byte-for-byte the same as before.
- pc convert already ran through the envelope get_wrapped() and the sandboxed
  render pipeline; verified it needs no in-process CAD.
- show() is unchanged.
- convert("build123d"/"cadquery") returns a live object and so still needs that
  library present; when it is missing it now warns naming the library and
  re-raises, instead of pointing at a 'partcad[cad]' install extra.
- drop the 'cad' optional-dependency entirely: remove requirements-cad.txt and the
  [cad] extra. The wheel now carries no CAD Requires-Dist, hard or optional.

Proven: with OCP/build123d/cadquery/ocpsvg blocked in the core process, 'pc import
assembly' writes all parts and 'pc convert' step->stl produces output - the
geometry work happens in sandbox subprocesses.

Validated: pytest partcad partcad-cli (443 passed, 21 skipped).
