--- partcad/requirements-cad.txt
@@ -1,10 +1,0 @@
-# Optional 'cad' extra: the CAD libraries the core needs only for its opt-in,
-# in-process features - convert("build123d"/"cadquery") and the deprecated
-# get_part_build123d/get_part_cadquery accessors, show() (the OCP CAD Viewer),
-# and 'pc import assembly' (the STEP-CAF reader). Everything else - building,
-# rendering and exporting parts, sketches and assemblies - happens in sandboxed
-# runtimes that install their own copies, so a plain 'pip install partcad' needs
-# none of these. Pinned to the versions the sandboxes use (see sandbox_versions).
-cadquery-ocp==7.9.3.1.1
-build123d==0.11.1
-ocpsvg==0.6.0
