+++ partcad/requirements.txt
@@ -1,10 +1,11 @@
-# No CAD library is a hard runtime dependency of the core any more. PartCAD
-# builds, renders and exports every shape in sandboxed runtimes (which install
-# their own copy - see sandbox_versions), and carries geometry between them as
-# BREP-byte envelopes rather than live OCP objects. cadquery-ocp, build123d and
-# ocpsvg are the optional 'cad' extra (see requirements-cad.txt), needed only by
-# the opt-in in-process features: convert("build123d"/"cadquery"), show() and
-# 'pc import assembly'.
+# No CAD library is a runtime dependency of PartCAD. It builds, renders and
+# exports every shape - and reads STEP assemblies for 'pc import assembly' and
+# converts files for 'pc convert' - in sandboxed runtimes, which install their
+# own copy of the CAD stack (see sandbox_versions), and it carries geometry
+# between them as BREP-byte envelopes rather than live OCP objects. The only
+# things that need a CAD library present in this process are handing back live
+# build123d/cadquery objects (convert("build123d"/"cadquery")) and show(); both
+# are opt-in, import their library lazily, and warn if it is not installed.
 # sdf-fork  # installed on demand into the sandbox runtime (see part_factory_sdf.py)
 pyyaml>=6.0.1
 pygit2>=1.15.0
